refactor(php): extract engine-agnostic PHP runtime to packages/adapter-php
One runtime, two engines (#2100, adapter-perl precedent): move the
engine-agnostic PHP server runtime (BarefootJS, Evaluator, SearchParams
+ the vector-corpus test suite) out of packages/adapter-twig/php into a
new shared packages/adapter-php (Composer: barefootjs/runtime, npm stub:
@barefootjs/php), so the upcoming Blade adapter can reuse it.

- Backend contract grows 4 -> 5 methods: template-variable-name mangling
  is now the backend's ident() (BarefootJS::render_child no longer calls
  twig_ident directly); TwigBackend::ident delegates to naming.php.
- Canonical JSON encoding moves to Barefoot\Json::canonicalEncode;
  TwigBackend::defaultJsonEncoder stays as a thin delegation.
- TwigBackend.php, naming.php and the live-render test stay in
  adapter-twig; its composer.json requires barefootjs/runtime via a
  path repository.
- Consumers updated: adapter-twig test-render, integrations/php
  (assemble-deps stages runtime + backend separately), docker-compose
  volume, ci-php.yml (split install/test steps), release.yml (new
  packagist-runtime-release split job; mirror repo provisioning is an
  infra follow-up), spec/vectors doc paths.

Verified: adapter-php runtime tests 412 pass, twig live render 9 pass,
bun test packages/adapter-twig 408 pass / 0 fail.
