+++ docker-compose.yml
@@ -249,6 +249,10 @@ services:
       # the workspace source dir directly (mirrors flask's sys.path
       # fallback) so local dev works without running assemble-deps first.
       - ./packages/adapter-twig/php:/workspace/packages/adapter-twig/php
+      # packages/adapter-twig/php/composer.json requires barefootjs/runtime
+      # via a `path` repository pointing at ../../adapter-php (a symlink in
+      # its vendor/ dir) -- that target must exist in the container too.
+      - ./packages/adapter-php:/workspace/packages/adapter-php
 
   site-core:
     build:
