+++ packages/adapter-php/package.json
@@ -1,0 +1,26 @@
+{
+  "name": "@barefootjs/php",
+  "version": "0.1.0",
+  "description": "BarefootJS engine-agnostic PHP runtime -- pluggable rendering backend; the shared basis for Twig, Blade and other PHP engine integrations",
+  "type": "module",
+  "files": [
+    "src",
+    "tests",
+    "composer.json",
+    "README.md"
+  ],
+  "keywords": [
+    "php",
+    "barefoot",
+    "ssr",
+    "runtime",
+    "template"
+  ],
+  "author": "kobaken <kentafly88@gmail.com>",
+  "license": "MIT",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/piconic-ai/barefootjs",
+    "directory": "packages/adapter-php"
+  }
+}
