+++ packages/adapter-php/src/Evaluator.php
