+++ packages/adapter-php/src/SearchParams.php
