+++ packages/adapter-php/tests/test_query.php
