+++ packages/adapter-twig/php/README.md
@@ -1,6 +1,18 @@
 # barefootjs/twig
 
-PHP runtime for the BarefootJS Twig adapter.
+Twig rendering backend (`Barefoot\TwigBackend`) for the BarefootJS PHP
+runtime.
+
+This package targets `twig/twig` and pairs with the `@barefootjs/twig`
+compile-time adapter, which emits `.twig` templates that call the runtime as
+a `bf` object (`{{ bf.scope_attr() }}`, `{{ bf.json(x) }}`,
+`{{ bf.spread_attrs(bag) }}`).
+
+The engine-agnostic runtime itself (`Barefoot\BarefootJS`, `Evaluator`,
+`SearchParams`, `Json`) lives in a separate package,
+[`packages/adapter-php`](../../adapter-php) (`barefootjs/runtime` on
+Composer) -- see that package's README for the backend contract every PHP
+engine adapter (Twig, Blade, ...) implements.
 
 This package is maintained in the BarefootJS monorepo and is mirrored to a
 Packagist-facing repository only for Composer distribution.
@@ -8,7 +20,23 @@ Packagist-facing repository only for Composer distribution.
 - Source of truth: <https://github.com/piconic-ai/barefootjs/tree/main/packages/adapter-twig/php>
 - Monorepo: <https://github.com/piconic-ai/barefootjs>
 - npm adapter: `@barefootjs/twig`
+- Runtime dependency: `barefootjs/runtime` (`packages/adapter-php`)
 
 Do not send implementation pull requests to the Packagist mirror. Send changes to
 the monorepo path above; the release workflow splits this directory and pushes the
 mirror automatically.
+
+## Tests
+
+```sh
+php tests/test_render.php
+```
+
+A live end-to-end Twig render (scope markers, hydration attrs, text slots,
+autoescaping, `spread_attrs`, reserved-word mangling). Needs `twig/twig`
+installed (`composer install`); skips gracefully with a notice (not a
+failure) when `vendor/autoload.php` is absent. The engine-agnostic runtime's
+own test suite (golden helper vectors, evaluator vectors, `spread_attrs`,
+`omit`, `render_child`, ...) lives in
+[`packages/adapter-php/tests`](../../adapter-php/tests) — run with
+`php packages/adapter-php/tests/run.php`.
